New Task with assigned users in select

+++ routes/tasks.js
@@ -12,10 +12,22 @@ export default (router) => {
       }
 
       console.log('in Get /tasks');
-      const tasks = await models.Task.findAll({ include: ['status', 'creator', 'tags'] });
-
+      const tasks = await models.Task.findAll({ include: ['assigned', 'status', 'creator'] });
+      console.log('tasks: ', tasks);
       ctx.render('tasks', { tasks });
-    })
+    }) // форма для создания новой задачи
+    .get('newTask', '/tasks/new', async (ctx) => {
+      console.log('In newTask');
+      const userIdsession = ctx.session.userId;
+      if (!userIdsession) {
+        ctx.flash.set('You need to autenticate!');
+        ctx.redirect('/');
+        return;
+      }
+      const task = models.Task.build();
+      const users = await models.User.findAll();
+      ctx.render('tasks/new', { f: buildFormObj(task), data: users });
+    }) // форма просмотра задачи
     .get('task', '/tasks/:id', async (ctx) => {
       console.log('In Task');
       const { id: taskId } = ctx.params;
@@ -30,17 +42,6 @@ export default (router) => {
         console.log('Error task findOne: ', e);
       }
       ctx.render('tasks/task', { f: buildFormObj(task) });
-    }) // форма для создания новой задачи
-    .get('newTask', '/tasks/new', async (ctx) => {
-      console.log('In newTask');
-      const userIdsession = ctx.session.userId;
-      if (!userIdsession) {
-        ctx.flash.set('You need to autenticate!');
-        ctx.redirect('/');
-        return;
-      }
-      const task = models.Task.build();
-      ctx.render('tasks/new', { f: buildFormObj(task) });
     }) // создание новой задачи
     .post('tasks', '/tasks', async (ctx) => {
       const userIdsession = ctx.session.userId;
@@ -50,7 +51,8 @@ export default (router) => {
         return;
       }
       const { request: { body: { form } } } = ctx;
-      console.log(`body data: ${JSON.stringify(form)}`);
+      console.log(`form data: ${JSON.stringify(form)}`);
+      // console.log('body data: ', JSON.stringify(ctx.request.body));
 
       console.log('tagsForm: ', form.tags);
       const regComma = /\s*,\s*/;
@@ -58,13 +60,16 @@ export default (router) => {
       console.log('tagsName: ', tagsName);
       const task = models.Task.build(form);
 
-      const user = await models.User.findOne({ where: { id: userIdsession } });
+      const userCreator = await models.User.findOne({ where: { id: userIdsession } });
       try {
-        task.setCreator(user);
+        task.setCreator(userCreator);
         console.log('Task build: ', JSON.stringify(task));
       } catch (e) {
         console.log('Set creator error: ', e);
       }
+
+      const userAssigned = await models.User.findOne({ where: { id: form.assigned } });
+      task.setAssigned(userAssigned);
 
       const [statusNew, created] = await models.TaskStatus.findOrCreate({ where: { name: 'New' } });
       console.log('Satus new created: ', created);
@@ -126,30 +131,34 @@ export default (router) => {
       }
       ctx.render('tasks/edit', { f: buildFormObj(task.get({ plain: true })) });
     }) // редактирование задачи
-    .patch('editTaskPatch', '/tasks/:id', async (ctx) => {
+    // .patch('editTaskPatch', '/tasks/:id', async (ctx) => {
 
-    }) // удаление задачи
+    // }) // удаление задачи
     .delete('deleteTask', '/tasks/:id', async (ctx) => {
+      const userIdsession = ctx.session.userId;
       // console.log('Session userId: ', userIdsession);
-      // const userIdsession = ctx.session.userId;
       const { id: taskId } = ctx.params;
-      // if (!userIdsession || userIdsession.toString() !== userId) {
-      //   ctx.flash.set('You need to autenticate!');
-      //   ctx.redirect('/');
-      //   return;
-      // }
 
       const task = await models.Task.findOne({
-        where: {
-          id: taskId,
-        },
+        where: { id: taskId },
+        include: ['creator'],
       });
 
-      await models.Task.destroy({
-        where: {
-          id: task.id,
-        },
-      });
+      // console.log('Creator id: ', task.creator.id, ' type: ', typeof task.creator.id);
+      if (!userIdsession || userIdsession.toString() !== task.creator.id.toString()) {
+        ctx.flash.set('You are not autorized to remove this task!');
+        ctx.redirect(router.url('tasks'));
+        return;
+      }
+      try {
+        await models.Task.destroy({
+          where: {
+            id: task.id,
+          },
+        });
+      } catch (e) {
+        console.log(e);
+      }
 
       ctx.redirect(router.url('tasks'));
     });
